Izbacio neke castove koji prave problem, promenio neke tipove parametara iz int u float, dodao direction kao parametar u metodi koja pravi listu igraca i namestio ga da bude random ugao od 0 do 2pi

+++ core/src/com/jrti/curveparty/Player.java
@@ -37,7 +37,7 @@ public interface Player {
 
     void move(); //todo v. NetworkPlayer, ovo u principu ne pripada ovde
 
-    void moveTo(int x, int y, int thickness);
+    void moveTo(float x, float y, int thickness);
 
     void setDirection(double direction);
 }
